refac: move agent.py to agent/ dir (separate files)
