+++ core/agent/learning.py
@@ -1,0 +1,20 @@
+from typing import Callable
+
+from core.agent.base import Agent
+from core.base_types import ObsType, ActType
+
+
+class ModelAgentSket(Agent):
+    def __init__(self):
+        # TODO: need joint channels info for the system:
+        #  it must know about its own presence
+        self.model: Callable[[ObsType], ActType] = None
+
+    def forward(self, obs: ObsType) -> ActType:
+        agents, medium = obs
+
+        action = self.model(obs)
+
+        result_action = self.postprocess_action(agents, action)
+
+        return result_action
